Split packages between game and engine
