--- adventure/koerier/runText.ts
@@ -1,6 +1,0 @@
-import "./locaties";
-import { convertGame } from "../dsl/dsl";
-import { runGame } from "../cli-client/run";
-import { Game } from "./game";
-
-convertGame<Game>(runGame);
