+++ adventure/src/dsl/world-types.ts
