+++ adventure/src/index.ts
@@ -1,0 +1,2 @@
+export { world, convertGame } from "./dsl/dsl";
+export { runGame as runCLIGame } from "./cli-client/run";
