+++ koerier/src/runText.ts
@@ -1,0 +1,5 @@
+import "./locaties";
+import { convertGame, runCLIGame } from "point-n-click";
+import { Game } from "./game";
+
+convertGame<Game>(runCLIGame);
